+++ admin/src/components/footer.js
@@ -49,7 +49,7 @@ export default function FooterComponent() {
             </a>
           </li>
         </ul>
-        <p>Tous droits réservés - Ministère de l&apos;éducation nationale, de la jeunesse et des sports - {new Date().getFullYear()}</p>
+        <p>Tous droits réservés - Ministère de l&apos;éducation nationale et de la jeunesse - {new Date().getFullYear()}</p>
         <ul>
           <li>
             <a href="https://www.gouvernement.fr/" target="_blank" rel="noreferrer">