+++ api/src/controllers/zammood.js
@@ -59,17 +59,21 @@ router.post("/ticket", passport.authenticate(["referent", "young"], { session: f
       subject: req.body.subject,
       message: req.body.message,
       fromPage: req.body.fromPage,
+      formSubjectStep1: req.body.subjectStep1,
+      formSubjectStep2: req.body.subjectStep2,
     };
     const { error, value } = Joi.object({
       subject: Joi.string().required(),
       message: Joi.string().required(),
       fromPage: Joi.string(),
+      formSubjectStep1: Joi.string(),
+      formSubjectStep2: Joi.string(),
     })
       .unknown()
       .validate(obj);
     if (error) return res.status(400).send({ ok: false, code: ERRORS.INVALID_PARAMS });
 
-    const { subject, message } = value;
+    const { subject, message, formSubjectStep1, formSubjectStep2 } = value;
     const userAttributes = await getUserAttributes(req.user);
     const response = await zammood.api("/v0/message", {
       method: "POST",
@@ -81,6 +85,8 @@ router.post("/ticket", passport.authenticate(["referent", "young"], { session: f
         firstName: req.user.firstName,
         lastName: req.user.lastName,
         source: "PLATFORM",
+        formSubjectStep1,
+        formSubjectStep2,
         attributes: [...userAttributes, { name: "page précédente", value: value.fromPage }],
       }),
     });