+++ app/src/scenes/inscription/Create/stepProfil/EduconnectZone.js
@@ -2,8 +2,8 @@ import React from "react";
 import { Col } from "reactstrap";
 import InfoIcon from "../../../../assets/InfoIcon2";
 import NouveauBadge from "../../../../assets/NouveauBadge";
-import FormRow from "../../../../components/form/FormRow";
 import EduConnectButton from "../../../../components/buttons/EduConnectButton";
+import FormRow from "../../../../components/form/FormRow";
 
 export function EduconnectZone({ connected }) {
   return (
@@ -11,23 +11,23 @@ export function EduconnectZone({ connected }) {
       <FormRow>
         <Col>
           {!connected ? (
-            <div className="flex justify-between flex-wrap gap-y-2">
-              <div className="flex items-center gap-4">
+            <div className="flex justify-center lg:justify-between flex-wrap gap-y-2">
+              <div className="flex flex-wrap items-center justify-center gap-x-4">
                 <NouveauBadge />
-                <div>
-                  <span className="inline-flex">
-                    Inscription avec Inscription avec ÉduConnect
-                    <a className="pl-1 flex items-center" href="https://educonnect.education.gouv.fr/educt-aide/informations/">
+                <div className="text-center">
+                  <span className="inline-flex items-center justify-center">
+                    Inscription avec ÉduConnect
+                    <a className="pl-1" href="https://educonnect.education.gouv.fr/educt-aide/informations/">
                       <InfoIcon />
                     </a>
                   </span>
-                  <p className="text-gray-500 text-xs font-weight-light">Un compte unique pour les services numériques des écoles et des établissements.</p>
+                  <p className="text-center text-gray-500 text-xs font-weight-light">Un compte unique pour les services numériques des écoles et des établissements.</p>
                 </div>
               </div>
               <EduConnectButton />
             </div>
           ) : (
-            <i>Les information en provenance de ÉduConnect ont bien été enregistrées.</i>
+            <i>Les informations en provenance de ÉduConnect ont bien été enregistrées.</i>
           )}
         </Col>
       </FormRow>