+++ app/src/scenes/phase2v2/desktop/components/CardEquivalence.js
@@ -7,6 +7,7 @@ import SimpleFileIcon from "../../../../assets/icons/SimpleFileIcon";
 import { copyToClipboard, formatDateFR, translateEquivalenceStatus } from "../../../../utils";
 import ModalFiles from "./ModalFiles";
 import { useHistory } from "react-router-dom";
+import Pencil from "../../../../assets/icons/Pencil";
 
 export default function CardEquivalence({ equivalence, young }) {
   const [open, setOpen] = React.useState(false);
@@ -53,24 +54,13 @@ export default function CardEquivalence({ equivalence, young }) {
       </div>
       {open ? (
         <>
-          {equivalence.status === "WAITING_VERIFICATION" ? (
-            <>
-              <hr className="border-gray-200 mb-3" />
-              <div className="flex justify-end">
-                <button
-                  className="mr-8 mb-3 border-[1px] border-indigo-600 hover:bg-indigo-600 text-indigo-600 hover:text-white px-4 py-2 rounded-lg"
-                  onClick={() => history.push(`/phase2/equivalence/${equivalence._id}`)}>
-                  Modifier ma demande
-                </button>
-              </div>
-            </>
-          ) : null}
+          <hr className="mb-3 text-gray-200" />
           {equivalence.status === "WAITING_CORRECTION" ? (
             <>
               <div className="flex justify-between items-center px-2 py-3 rounded-lg bg-gray-50 mb-4 gap-6">
                 <div className="flex flex-col flex-1">
                   <div className="text-base font-semibold">Corrections demandées</div>
-                  <div className="text-sm text-gray-500">{equivalence.correctionMessage}</div>
+                  <div className="text-sm text-gray-500">{equivalence.message}</div>
                 </div>
                 <button
                   className="mr-4 border-[1px] border-indigo-600 hover:bg-indigo-600 text-indigo-600 hover:text-white px-4 py-2 rounded-lg"
@@ -85,68 +75,81 @@ export default function CardEquivalence({ equivalence, young }) {
               <div className="flex justify-between items-center px-2 py-3 rounded-lg bg-gray-50 mb-4 gap-6">
                 <div className="flex flex-col flex-1">
                   <div className="text-base font-semibold">Motif du refus</div>
-                  <div className="text-sm text-gray-500 w-3/4">{equivalence.refusedMessage}</div>
+                  <div className="text-sm text-gray-500 w-3/4">{equivalence.message}</div>
                 </div>
               </div>
             </>
           ) : null}
           <div className="flex items-stretch mb-3 gap-4 justify-around">
-            <div className="grid grid-cols-2 py-2">
-              <div className="flex flex-col gap-y-4 text-sm leading-none font-normal text-gray-400">
-                <span>Type d’engagement :</span>
-                <span>Structure d’accueil :</span>
-                <span>Dates :</span>
-                {equivalence.frequency ? <span>Fréquence :</span> : null}
-                <span>Adresse :</span>
-                <span>Code postal :</span>
-                <span>Ville :</span>
-              </div>
-              <div className="flex flex-col gap-y-4 text-sm leading-none font-medium">
-                <span>{equivalence.type}</span>
-                <span>{equivalence.structureName}</span>
-                <span>
-                  Du {formatDateFR(equivalence.startDate)} au {formatDateFR(equivalence.endDate)}
-                </span>
-                {equivalence.frequency ? (
-                  <span className="lowercase">
-                    {equivalence.frequency.nombre} {equivalence.frequency.duree} {equivalence.frequency.frequence}
+            <div className="flex ">
+              <div className="grid grid-cols-2 py-2 items-center">
+                <div className="flex flex-col gap-y-4 text-sm leading-none font-normal text-gray-400">
+                  <span>Type d’engagement :</span>
+                  <span>Structure d’accueil :</span>
+                  <span>Dates :</span>
+                  {equivalence.frequency ? <span>Fréquence :</span> : null}
+                  <span>Adresse :</span>
+                  <span>Code postal :</span>
+                  <span>Ville :</span>
+                </div>
+                <div className="flex flex-col gap-y-4 text-sm leading-none font-medium">
+                  <span>{equivalence.type}</span>
+                  <span>{equivalence.structureName}</span>
+                  <span>
+                    Du {formatDateFR(equivalence.startDate)} au {formatDateFR(equivalence.endDate)}
                   </span>
-                ) : null}
-                <span>{equivalence.address}</span>
-                <span>{equivalence.zip}</span>
-                <span>{equivalence.city}</span>
-              </div>
-            </div>
-            <div className="flex flex-col justify-center bg-gray-50 rounded-lg gap-4">
-              <div className="flex flex-col justify-center items-center gap-2 mx-16">
-                <SimpleFileIcon />
-                <div className="text-sm leading-5 font-bold text-center">
-                  Document justificatif <br /> d’engagement
+                  {equivalence.frequency ? (
+                    <span className="lowercase">
+                      {equivalence.frequency.nombre} {equivalence.frequency.duree} {equivalence.frequency.frequence}
+                    </span>
+                  ) : null}
+                  <span>{equivalence.address}</span>
+                  <span>{equivalence.zip}</span>
+                  <span>{equivalence.city}</span>
                 </div>
               </div>
-              <div className="flex flex-col justify-end items-end px-7">
-                <div className="transition duration-150 flex rounded-full bg-blue-600 p-2 items-center justify-center hover:scale-110 ease-out hover:ease-in cursor-pointer">
-                  <Download className=" text-indigo-100 bg-blue-600 " onClick={() => setModalFiles({ isOpen: true })} />
+              {equivalence.status === "WAITING_VERIFICATION" ? (
+                <div className="flex items-end justify-start">
+                  <button
+                    className="flex items-center justify-center rounded-full bg-blue-100 p-2 hover:scale-105 cursor-pointer mb-2"
+                    onClick={() => history.push(`/phase2/equivalence/${equivalence._id}`)}>
+                    <Pencil className="text-blue-600" />
+                  </button>
+                </div>
+              ) : null}
+            </div>
+            <div className="flex items-stretch gap-6">
+              <div className="flex flex-col justify-center bg-gray-50 rounded-lg gap-4">
+                <div className="flex flex-col justify-center items-center gap-2 mx-16">
+                  <SimpleFileIcon />
+                  <div className="text-sm leading-5 font-bold text-center">
+                    Document justificatif <br /> d’engagement
+                  </div>
+                </div>
+                <div className="flex flex-col justify-end items-end px-7">
+                  <div className="transition duration-150 flex rounded-full bg-blue-600 p-2 items-center justify-center hover:scale-110 ease-out hover:ease-in cursor-pointer">
+                    <Download className=" text-indigo-100 bg-blue-600 " onClick={() => setModalFiles({ isOpen: true })} />
+                  </div>
                 </div>
               </div>
-            </div>
-            <div className="flex flex-col justify-center border-[1px] border-gray-200 rounded-lg py-4 px-8">
-              <div className="text-base leading-6 font-bold text-gray-900 mb-4">
-                Personne contact au sein <br /> de la structure
-              </div>
-              <div className={`h-10 w-10 flex justify-center items-center rounded-full bg-gray-100 text-indigo-600 text-base font-semibold mb-3`}>
-                {getInitials(equivalence.contactFullName)}
-              </div>
-              <div className="text-sm leading-5 font-medium text-gray-900 mb-2">{equivalence.contactFullName}</div>
-              <div className="flex items-center mb-4">
-                <div className="text-xs leading-none font-nornal text-fray-700 mr-2 ">{equivalence.contactEmail}</div>
-                <div
-                  className="flex items-center justify-center cursor-pointer hover:scale-105"
-                  onClick={() => {
-                    copyToClipboard(equivalence.contactEmail);
-                    setCopied(true);
-                  }}>
-                  {copied ? <HiCheckCircle className="text-green-500" /> : <BiCopy className="text-gray-400" />}
+              <div className="flex flex-col justify-center border-[1px] border-gray-200 rounded-lg py-4 px-8">
+                <div className="text-base leading-6 font-bold text-gray-900 mb-4">
+                  Personne contact au sein <br /> de la structure
+                </div>
+                <div className={`h-10 w-10 flex justify-center items-center rounded-full bg-gray-100 text-indigo-600 text-base font-semibold mb-3`}>
+                  {getInitials(equivalence.contactFullName)}
+                </div>
+                <div className="text-sm leading-5 font-medium text-gray-900 mb-2">{equivalence.contactFullName}</div>
+                <div className="flex items-center mb-4">
+                  <div className="text-xs leading-none font-nornal text-fray-700 mr-2 ">{equivalence.contactEmail}</div>
+                  <div
+                    className="flex items-center justify-center cursor-pointer hover:scale-105"
+                    onClick={() => {
+                      copyToClipboard(equivalence.contactEmail);
+                      setCopied(true);
+                    }}>
+                    {copied ? <HiCheckCircle className="text-green-500" /> : <BiCopy className="text-gray-400" />}
+                  </div>
                 </div>
               </div>
             </div>