+++ app/src/scenes/phase2v2/views/EditEquivalence.js
@@ -9,7 +9,6 @@ import ChevronDown from "../../../assets/icons/ChevronDown";
 import InformationCircle from "../../../assets/icons/InformationCircle";
 import PaperClip from "../../../assets/icons/PaperClip";
 import api from "../../../services/api";
-import { RiErrorWarningLine } from "react-icons/ri";
 
 export default function EditEquivalence() {
   const young = useSelector((state) => state.Auth.young);
@@ -89,7 +88,7 @@ export default function EditEquivalence() {
         const { ok, data } = await api.get(`/young/${young._id.toString()}/phase2/equivalence/${equivalenceId}`);
         if (ok) {
           setData(data);
-          if (data?.frequency.nombre && data?.frequency.duree && data?.frequency.frequence) setFrequence(true);
+          if (data?.frequency?.nombre && data?.frequency?.duree && data?.frequency?.frequence) setFrequence(true);
           return;
         }
       })();
@@ -125,6 +124,7 @@ export default function EditEquivalence() {
         delete data.youngId;
         delete data.createdAt;
         delete data.__v;
+        data.status = "WAITING_VERIFICATION";
         const { ok } = await api.put(`/young/${young._id.toString()}/phase2/equivalence/${equivalenceId}`, data);
         if (!ok) {
           toastr.error("Oups, une erreur est survenue");
@@ -155,13 +155,10 @@ export default function EditEquivalence() {
           </div>
         </div>
         {data.status === "WAITING_CORRECTION" ? (
-          <div className="border-[1px] border-[#FD7A02] rounded-lg bg-[#FD7A02]/10 mt-4">
-            <div className="flex items-center px-4 py-3">
-              <RiErrorWarningLine className="text-[#FD7A02]" />
-              <div className="flex flex-col flex-1">
-                <div className="ml-4 text-[#FD7A02] text-sm leading-5 font-medium">Corrections demandées</div>
-                <div className="ml-4 text-[#FD7A02] text-sm leading-5 font-medium">{data?.correctionMessage}</div>
-              </div>
+          <div className="border-[1px] border-gray-200 rounded-lg bg-white mt-4">
+            <div className="flex flex-col px-3 py-2 gap-2">
+              <div className="text-neutral-900 text-base leading-5 font-medium">Corrections demandées</div>
+              <div className="text-gray-500 text-sm leading-5 font-medium">{data?.message}</div>
             </div>
           </div>
         ) : null}
@@ -271,6 +268,7 @@ export default function EditEquivalence() {
               <input
                 className="w-full text-sm leading-5 font-normal ::placeholder:text-gray-500"
                 placeholder="Date de fin"
+                min={formatDate(data?.startDate)}
                 type="date"
                 value={formatDate(data?.endDate)}
                 onChange={(e) => setData({ ...data, endDate: e.target.value })}
@@ -447,7 +445,7 @@ export default function EditEquivalence() {
           className="rounded-lg w-full py-2 mt-4 text-sm leading-5 font-medium bg-blue-600 text-white border-[1px] border-blue-600 hover:bg-white hover:!text-blue-600 disabled:bg-blue-300 disabled:!text-white disabled:border-blue-300"
           disabled={loading || uploading}
           onClick={() => handleSubmit()}>
-          {loading ? "Chargement" : "Valider ma demande"}
+          {loading ? "Chargement" : "Modifier ma demande"}
         </button>
       </div>
     </div>